環境モニタリング押下後LeftPlltabのホバーが点滅するバグを解消
